feat(*): updating implementation according to feedbacks
