+++ packages/manager/apps/container/src/helpers/piano-override.ts
@@ -1,0 +1,11 @@
+// Piano Analytics override
+export {};
+
+declare global {
+    interface Window {
+      _pac: any;
+    }
+  }
+window._pac = window._pac || {};
+window._pac.cookieSecure = "true";
+window._pac.cookieSameSite = "None";
